refactor(#3479): extract launch parsing from tui_prompt_relay.rs
Move the Claude TUI launch-script parsing cluster verbatim out of
tui_prompt_relay.rs into a new sibling submodule
src/services/discord/tui_prompt_relay/launch_script.rs.

Moved (behavior-preserving, #[cfg(unix)]):
- struct ClaudeTuiLaunchInfo
- parse_claude_tui_launch_script
- parse_claude_tui_launch_script_content
- shell_words_from_line
- test parses_claude_tui_launch_script_content

Bodies are byte-identical; only visibility (pub(super)) and module
wiring changed. parse_claude_tui_launch_script is re-imported so the
parent and the rehydration sibling resolve it unchanged.

tui_prompt_relay.rs giant-file ratchet: 4458 -> 4376 prod LoC.
Synced change-surfaces.md freeze entry and regenerated inventory docs.
