+++ src/services/discord/tui_prompt_relay.rs
@@ -91,6 +91,19 @@ pub(in crate::services::discord) use self::anchor_completion::{
     should_complete_tui_direct_anchor_lifecycle,
 };
 
+// #3479: the Claude TUI launch-*script* parsing cluster (`ClaudeTuiLaunchInfo`,
+// the file/content parsers, and the single-quote shell-word splitter) moved
+// verbatim to a capped sibling module. Every item is unix-only, so the module
+// decl and the re-import are `#[cfg(unix)]`-gated. Only
+// `parse_claude_tui_launch_script` is reached from outside the child (the parent
+// `claude_tui_launch_context` caller and, via this parent's `use super::*;`
+// glob, the sibling `rehydration` module), so it alone is re-imported here to
+// keep those call sites byte-identical.
+#[cfg(unix)]
+mod launch_script;
+#[cfg(unix)]
+use self::launch_script::parse_claude_tui_launch_script;
+
 const SSH_DIRECT_PROMPT_PREVIEW_LIMIT: usize = 1500;
 const CODEX_IDLE_ROLLOUT_POLL_INTERVAL: Duration = Duration::from_millis(500);
 const CLAUDE_IDLE_REHYDRATE_POLL_INTERVAL: Duration = Duration::from_secs(5);
@@ -3080,101 +3093,6 @@ fn advance_claude_tmux_runtime_binding_offset(
         transcript_path.to_str().unwrap_or_default(),
         offset,
     )
-}
-
-#[cfg(unix)]
-#[derive(Debug, PartialEq, Eq)]
-struct ClaudeTuiLaunchInfo {
-    working_dir: PathBuf,
-    session_id: String,
-}
-
-#[cfg(unix)]
-fn parse_claude_tui_launch_script(path: &Path) -> Result<ClaudeTuiLaunchInfo, String> {
-    let script = std::fs::read_to_string(path)
-        .map_err(|error| format!("read Claude TUI launch script {}: {error}", path.display()))?;
-    parse_claude_tui_launch_script_content(&script)
-        .ok_or_else(|| format!("parse Claude TUI launch script {}", path.display()))
-}
-
-#[cfg(unix)]
-fn parse_claude_tui_launch_script_content(script: &str) -> Option<ClaudeTuiLaunchInfo> {
-    let mut working_dir: Option<PathBuf> = None;
-    let mut session_id: Option<String> = None;
-    for line in script.lines() {
-        let words = shell_words_from_line(line.trim());
-        if words.first().is_some_and(|word| word == "cd") {
-            if let Some(dir) = words.get(1).filter(|value| !value.trim().is_empty()) {
-                working_dir = Some(PathBuf::from(dir));
-            }
-            continue;
-        }
-        if !words.first().is_some_and(|word| word == "exec") {
-            continue;
-        }
-        for pair in words.windows(2) {
-            if matches!(pair[0].as_str(), "--session-id" | "--resume") && !pair[1].trim().is_empty()
-            {
-                session_id = Some(pair[1].clone());
-                break;
-            }
-        }
-    }
-    Some(ClaudeTuiLaunchInfo {
-        working_dir: working_dir?,
-        session_id: session_id?,
-    })
-}
-
-#[cfg(unix)]
-fn shell_words_from_line(line: &str) -> Vec<String> {
-    let mut words = Vec::new();
-    let mut current = String::new();
-    let mut saw_word = false;
-    let mut in_single = false;
-    let mut chars = line.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        if in_single {
-            if ch == '\'' {
-                in_single = false;
-            } else {
-                current.push(ch);
-            }
-            saw_word = true;
-            continue;
-        }
-
-        if ch.is_whitespace() {
-            if saw_word {
-                words.push(std::mem::take(&mut current));
-                saw_word = false;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' => {
-                in_single = true;
-                saw_word = true;
-            }
-            '\\' => {
-                if let Some(next) = chars.next() {
-                    current.push(next);
-                    saw_word = true;
-                }
-            }
-            _ => {
-                current.push(ch);
-                saw_word = true;
-            }
-        }
-    }
-
-    if saw_word {
-        words.push(current);
-    }
-    words
 }
 
 #[cfg(unix)]
@@ -6930,24 +6848,6 @@ mod tests {
         assert!(
             gateway.can_chain_locally(),
             "bridge adapter still owns Discord delivery for the already-submitted turn"
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parses_claude_tui_launch_script_content() {
-        let script = concat!(
-            "#!/bin/bash\n",
-            "cd '/tmp/project'\\''s dir'\n",
-            "exec '/usr/local/bin/claude' '--dangerously-skip-permissions' '--session-id' '01234567-89ab-cdef-0123-456789abcdef' '--settings' '/tmp/settings.json'\n",
-        );
-
-        assert_eq!(
-            parse_claude_tui_launch_script_content(script),
-            Some(ClaudeTuiLaunchInfo {
-                working_dir: PathBuf::from("/tmp/project's dir"),
-                session_id: "01234567-89ab-cdef-0123-456789abcdef".to_string(),
-            })
         );
     }
